% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_file_size.R
\name{get_one_file_size}
\alias{get_one_file_size}
\title{Obtenir la taille d'un fichier}
\usage{
get_one_file_size(path)
}
\arguments{
\item{path}{Chemin vers le fichier}
}
\value{
df Un dataframe contenant le nom et la taille du fichier
}
\description{
Obtenir la taille d'un fichier
}
\examples{
get_one_file_size(system.file("squirrels_hex.png",package = "toolboxfileaxel"))

}
